Update age.R default age range
The default values before were 20 to 35.  The default values are buried in the documentation, so it took me a bit of wandering around to find out how and where they were set.  If there is not a specific reason to select 20 to 35 as the age range, I suggest expanding the default range to 18 to 89.  That is equally arbitrary, but it does have the advantage of matching the categories from the GSS.  That change was necessary for my own use case, and I think in general users may be expecting a bit of a larger range of ages here.